Add a basic scrolling terrain viewing widget.
The terrain view is essentially a container for a QGraphicsScene and QGraphicsView (check documentation).
It will generate a grid of QPixmap tiles, based on a given input.
Users can use scrollbars to pan around the terrain.  Hopefully we can get some kind of zoom in-out function working.

The interface is still under development.  It's constructor will probably at least take a TileMap<size_t>,
and render the appropriate tiles based on the index.  Maybe there are other things we'd want to add.

A test is currently hard-coded within the constructor, that will be replaced.
The next commit will be a gradient of noise, rather than by the random numbers of the current test.